Add parameter discard_guide_positions_overhanging_amplicon_edge
If run with param -- discard_guide_positions_overhanging_amplicon_edge, for guides that align to multiple positions, guide positions will be discarded if plotting around those regions would included bp that extend beyond the end of the amplicon. Normally this would cause CRISPResso to fail if plotting were requested beyond the end of an amplicon.