Fix a bug in `find_indels_substitutions`
The bug that this commit fixes is when an insertion occurs at the edge of the
include indexes. The trouble with this earlier was that it was using the `idx`
to calculate the size of the insertion, but the `idx` wasn't being incremented
anymore because it was outside of the include window.