Fix bug where plots were not filling up entire modal.
I have tried countless different ways for this to work, and this is the best
that I can come up with. After the modal is opened it triggers the plot to
resize, and then for some reason you need to trigger the resize event. I think
this is because a `div` changing size won't actually trigger the resizing of the
plot (and neither will just calling `Plotly.Plots.resize`...?!).