Abandon attempts to make plots fullscreen
This includes removing the Bootstrap modal (two out of the three plots would
resize properly and I couldn't figure out a way to have the plot displayed
outside of the modal). I have left in some javascript to make the plot
fullscreen, but I couldn't get the formatting quite right and the plot wasn't
much bigger in the fullscreen version because there was a ton of space between
the plot and the heatmap. If some brave soul would like to tackle it, feel free!