Don't run global frameshift plot when there are no reads
When there are no reads (i.e. global_MODIFIED_FRAMESHIFT +
global_MODIFIED_NON_FRAMESHIFT + global_NON_MODIFIED_NON_FRAMESHIFT == 0) there
was a bug when trying to compute the pie chart, because all of the values in the
pie chart are 0. This fix, will make sure that there is at least one read in
order for the plot to bee constructed properly.