Add StatusHandler to CRISPRessoCORE log
This will take the latest log output and write it to a file (`status.txt`), the
catch being that with each log the file is overwritten so that one can easily
tell where CRISPResso currently is and what the error is (if any). These changes
include some slight refactoring in order to accomodate any potential parameter
exceptions.