Status Updates + Pooled Mixed Mode Update (#279)
* Implement logging handler to overwrite the latest log status to file

* Add StatusHandler to CRISPRessoCORE log

This will take the latest log output and write it to a file (`status.txt`), the
catch being that with each log the file is overwritten so that one can easily
tell where CRISPResso currently is and what the error is (if any). These changes
include some slight refactoring in order to accomodate any potential parameter
exceptions.

* Add StatusHandler to CRISPRessoBatch and refactor `logger.warn` to `warn`

* Add StatusHandler to CRISPRessoPooled and a little refactoring

* Implement `percent_complete` to the status log

* Add StatusHandler to CRISPRessoAggregate log

* Add StatusHandler to CRISPRessoCompare log

* Add StatusHandler to CRISPRessoPooledWGSCompare log

* Add StatusHandler to CRISPRessoWGS log

* Rename `status.txt` to `CRISPResso_status.txt`

* Modify status log names to match the tool they are generated from

* Add percent_complete stages to CRISPRessoCORE

These also include log statements of each plot that is being generated as well
as fixing some variable name collisions with `ind`.

* Format the percentage in the log to be 2 decimal places

* Change all plotting logs from `info` to `debug` and simplify progress

This refactors how the progress of the plots is calculated, making it much
simplier. Before this change we would of had to keep track of the number of
times `percent_complete` was output, but now it simply updates the percent
complete after each amplicon is finished processing. Hopefully this will make
things easier to mantain even though it will be a little less "accurate" (not
sure how accurate the original implementation was...).

* Implemented shared console log handler across all CRISPResso* calls

This allows for easy changes to logging formatting, which was inspired by having
to change the default logging level. The default logging level needs to be set
at `logging.DEBUG` in order for the debug log statements to not be ignored for
the running and status logs.

* Add ability to set the verbosity level to each CRISPResso* tool

This allows users to set a verbosity level between 1 and 4 using the
`-v`/`--verbosity` CLI parameter. If the `--debug` flag is present, then the
level will default to 4, being the most verbose.

* Implement showing the last seen `percent_compelte` when none is provided

* Keep track of and log when multiple parallel runs are completed

These changes modify `CRISPRessoMultiProcessing.run_crispresso_cmds` such that
we can now display when a run is completed. This potentially breaks how
signals and interupts are handled with multiple runs happening, but this needs
to be reviewed.

* Add debug and percentage complete to CRISPRessoBatch

* Add percent complete to CRISPRessoPooled

* Add debug and percent_complete message to CRISPRessoAggregate

* Add `percent_complete` to CRISPRessoCompare

* Add `percent_complete` to CRISPRessoPooledWGSCompare

* Add status and `percent_complete` to CRISPRessoMeta

* Add `verbosity` arguments to CRISPRessoCompare and CRISPRessoPooledWGSCompare

* Fixing documentation to match pooled headers

* Header removal bug fix change documentation to guide_seq

* Update documentation and help feature for CRISPRessoPooled

* Remove extra newlines from CRISPRessoPooled -h

* Make variable names as clear as my firstborn child's name

* Update one more variable name

* Fix bug to flow CRISPRessoPooled options to sub command

* Make amplicon file args variable name clear

* Update how parameters are set and retrieved from parameter object

The refactor in the previous commit changed the type of the arguments to a
dictionary which doesn't have the parameters as attributes, and this commit
fixes that error.

* Add note in output header for change in default CRISPRessoPooled

In the next release (2.3.0) the `--demultiplex_only_at_amplicons` will be the
default when running in mixed-mode. This is to allow for inexact alignments of
the reads and the amplicons to the genome. For more context, see this issue
https://github.com/pinellolab/CRISPResso2/issues/276

* Clarify the verbosity parameter help message

* Separate out parameters to `normalize_name` in CRISPRessoCORE

* Separate out parameters to `normalize_name` in CRISPRessoWGS

* Separate out parameters to `normalize_name` in CRISPRessoPooled

* Separate out parameters to `normalize_name` in CRISPRessoCompare

* Fix bug in CRISPRessoPooled by replacing `database_id` with `normalize_name`

* Refactor `run_crispresso_cmds` to not require a `logger`

This commit implements the functionality to make the `logger` object optional by
seeing which module called the `run_crispresso_cmds` function and obtaining the
correct object from that module name.

The function also immediately returns when no commands are passed to it.

* Add amplicon name to plotting debug statements in CRISPRessoCORE

---------

Co-authored-by: Cole Lyman <cole@coles-mbp-2.lan>
Co-authored-by: Cole Lyman <cole@Coles-MacBook-Pro-2.local>
Co-authored-by: Cole Lyman <cole@colelyman.com>
Co-authored-by: Samuel Nichols <Snic9004@gmail.com>