Fix deletion at second position (#131)
* Fix bug when there is a deletion starting at the second position

This bug only happens when a deletion starts are the second position, before the
fix, it would report that the deletion started at the first position. It is
fixed now, so deletions at the second position are reported correctly.

* Update CRISPRessoCOREResources.c due to change in .pyx

* Add tests for find_indels_substitutions for deletions at the end

* Fix 1bp deletions at the end, and off by one error

This ensures that when a deletion occurs at the end of a read, the entire
deletion is accounted for.

* Update CRISPRessoCOREResources.c to reflect fixes for deletions at the end of alignments

* Add extra asserts to deletion checks

* Point to new test branch

* Reafctor deletion_coordinates to go past the end of the string for deletions at
the end of the sequence

* Point tests to master